Windows-related fixes

+++ MEArec/tools.py
@@ -475,7 +475,7 @@ def recursively_save_dict_contents_to_group(h5file, path, dic):
         Path to the h5 field
     """
     for key, item in dic.items():
-        if isinstance(item, (int, float, np.int64, np.float64, str, bytes, np.bool_)):
+        if isinstance(item, (int, float, np.integer, np.float, str, bytes, np.bool_)):
             if isinstance(item, np.str_):
                 item = str(item)
             h5file[path + key] = item
